Inline headers from demo-loop
